PLUG-373 Adds option to use absolute URLs when requiring resources.
All WebResourceManager methods that construct URLs have been overloaded to accept an additional urlMode parameter.  This parameter specifies whether to use absolute URLs, use relative URLs, or let the implementation decide.
In order to implement this, a corresponding overloaded verison of getBaseUrl was added to WebResourceIntegration. Applications that use Atlassian Plugins 2.3 will need to implement this method.
Along with the new methods, the JavaDoc for WebResourceManager was expanded and clarified.

git-svn-id: https://studio.atlassian.com/svn/PLUG/trunk@40248 2c54a935-e501-0410-bc05-97a93f6bca70