PLUG-599 - I have reverted the implementation to be something more sane. This has the down-side of overriding a method in the test class but I whole-heartedly feel that it is better than what this implementation has become. I do not feel that we need four interfaces and six classes to perform this functionality. I found the implementation WAY too verbose and convoluted.
I completely agree that we want to separate the classpath file loading from the logic of using that information so I have left that separation. 
I would like to see this be the last commit as this simple change has gotten out of hand, but I would not be surprise if it is not.

git-svn-id: https://studio.atlassian.com/svn/PLUG/trunk@64991 2c54a935-e501-0410-bc05-97a93f6bca70